(GRIPE)
If we list verstub.obj in the dialog for extra dependencies, it counts it
as a full dep; since we change it every build, that means the binary
is never actually considered up to date.  We need to list it as an extra
option to the linker, so it gets linked without being factored into
dependencies.