Gaaah.  Head hurts.
Fix rounding error: when rect_width is 9, we want
rect_cx = pRect->left + 4.5, not 4.  This resulted in a gap between the
black quad and the gradient quad in TransitionFadeWipe, and possibly
other obscure things.  (The transition doesn't use a texture at all, so this
wasn't actually related to the bilinear filtering issue.)