Clapm the fast-moving wheel sound effects; bumps speed from 80fps to 130fps for me.  (I think it was starting over a hundred at a time ...)
The "random" effect was experimental.  I liked the bounce, but the
effect of pulling in random songs as it spun was hacky and not too
important (and a little buggy).

Random is now like roulette, except 1: it's shuffled, 2: it moves in reverse,
and 3: it stops with a bounce instead of shifts, and 4: it spins twice as
fast.  Currently, the bounce is a little bumpy; that's loading the sample
music, I think.