Changing StdAfx.* to global.*.
First, we have a case mismatch ("stdafx"), so we're stuck with either
a rename or a global search and replace anyway.

SF is weird about renaming files: if I try removing "StdAfx" and add
"stdafx", the CVS server will crash.

I'm renaming this to "global" for another reason: "stdafx" is a well-known
filename for Windows's special-case PCH; I don't want to confuse porters
who might ignore it.  (Besides, it has nothing to do with AFX ...)