Small movie texture optimization: honor the texture color depth.
Another biggie is that we're copying every frame an extra time in
MovieTexture_DShow::NewData().  This is rather hard to fix without
bringing back the stability problems we had before; I need to think
about how to do this.  (Commenting out this memcpy boosts my test
video to about 22fps, so this isn't minor.)