Override the "size" of a texture:  Foo [res 128x64].png
This can be used to store a texture at a different resolution without
having to change all of the metrics that reference it.

The most obvious use is to lower the resolution of large images.

A less obvious use is to raise the resolution of images.  If an image is
going to be zoomed, it may be useful to store the texture at a higher
resolution than native.

This can be done with metrics, but it's a pain to have to track down
each metric that needs to be changed.  This (should) also work with
fonts (but that needs testing).

Use this sparingly.