Load global Lua scripts.
This is appropriate for global scripts which are loaded only once, and
used multiple times: the scripts can safely affect global state (eg. override
functions), and stay in memory.  There's no way to unload them short
of resetting the Lua state.  It's not appropriate for one-shot scripts, such
as those that might be associated with BGAnimations.  (needs more work)