add RageTexturePreloader.
This obsoletes Cached and Permanent textures, implementing them more
simply: if you want to keep the texture loaded, just load it and keep the texture
around as long as you want it.  The old "texture policy" scheme of "keep the
texture around for the duration of the screen" was ambiguous--we can have any
number of screens loaded.