Create actors with ActorUtil::Create in ActorUtil::MakeActor.  This means ActorUtil::LoadFromNode always uses ActorUtil::Create.  Previously, sprites and models loaded without an explicit Class= were loaded directly.
Nested XML is merged explicitly.  Before, we loaded the final actor, and then
called Actor::LoadFromNode to pull in attributes from each nested XML node
as ActorUtil::LoadFromNode recursed back out, which was a bit confusing.  It
also caused weirdness with InitCommand: each time called it again, possibly
with a different InitCommand.  We need to only call it once, after all nodes
are loaded.

(MergeActorXML is a little annoying, but it does make it more clear what's
really happening.)