This code isn't worth it.  It's only used for one driver.  It's mostly useful for pointing casual users to drivers, but any URLs we give them will just go out of date, and 404s don't help people. It's also too heavyweight to be done in ArchHooks; aim to not use anything in it that we don't want to always include.
(I'm sort of undecided about this. This definitely needs to be out
of HOOKS; it could be moved to GraphicsWindow.  It's potentially
useful, in theory, but it'd be a major hassle to maintain and always
be incomplete.  Since it's only used for one ancient card and has
never been used for anything else, I'm just removing it for now.)