IPreference stores preferences, but knows nothing about loading or saving them; that's up to PrefsManager.
Treat IPreference as a simple data holder, with simple facilities for
looking them up by name, converting to/from strings, and storing to/from
an XNode, but with no application-specific code to save to disk.  This
can be used alone as a way for code to configure things internally, even
for uses that have no notion of storing user preferences (eg. unit tests),
and other programs can use it to store preferences in entirely different
ways (SMPackage could use it to store to the registry).  PrefsManager is
layered on top, to implement StepMania's particular use of Preference
(saving and loading INIs), but isn't needed for Preference to be useful.

This also makes Preference only use XNode, not the more specialized IniFile.

(One piece is missing: several low-level places, eg. Dialog, want to set
a preference and write it to disk immediately.  The only way to do that
is to have access to PREFSMAN.  FIXME.)