Split out tween weighting functions.  Add tweening from 1d and 2d bezier curves.
The "bouncebegin" and "bounceend" tweens are now implemented in Lua,
using 2D beziers which approximate the old behavior very closely.
Added "ease" tween, which works like Flash's.

Custom tweens can be used, passing in a table of 4 values for a
1D bezier, or 8 values for 2D.

   tween,.5,TweenBezier,{ 0, p1, p2, 1 }
   tween,.5,TweenBezier,{ 0, 0, x1, y1, x2, y2, 1, 1 }

As with all tweens, the first parameter is the duration.

The 1D bezier { a, b, c, d } is equivalent to the 2D bezier
{ 0, a, 1/3, b, 2/3, c, 1, d }, but a 1D bezier is more efficient.