oops.  Do keep the non-offset beat around, and use it for actor effects. (This should probably be used for more than this; this stuff needs a cleanup ...)
(test case: pick a song with a stop near the beginning of the sample music,
eg. Fly With Me; no matter what the offset is set to, it should always show
the same pause on beat effects.  This could do so while honoring the offset,
but getting that right in GameSoundManager is very tricky, and it's usually
not important for non-gameplay cosmetics.)