Replace PRNG with a standard MT.
RandomGen r; r(); now generates [0,2^31-1] on 32-bit platforms,
not [0,2^31-2].  No longer uses MAX_INT, so the default behavior
doesn't depend on the size of an int.

RandomInt no longer generates a float and truncates it.  That
limited the values that could be returned to those that fit in
a 32-bit float.

Remove NR code.  It's not free.

This code is not threadsafe.  Locking access would kill performance.
The old code wasn't, either.  I think most symptoms would be
innocuous enough to not worry about.