FlipHeadAndTailWhenReverse is a bit confusing.
Other things are happening here: anchoring the hold note to the
top or bottom, and flipping the wavy texture upside down.  These
have separate purposes:

Normal hold rendering anchored to the bottom, so the hold scrolls
up as it's held; if it was anchored at the top, then it would
stay in place and the bottom would disappear as it was held.
In reverse, we can optionally reverse this, so the hold scrolls
down as it's held, or not reverse it, causing the hold to be
eaten by the head as the head comes down.

Normal hold rendering displays the hold body right-side-up.
In reverse, we can optionally render the hold body upside-down.
If we're flipping the head and tail (usually, putting the
overlaid arrow on the bottom instead of the top), mirroring
the body may or may not be wanted.

Additionally, we can optionally reverse the head and tail
(not to be confused with the top and bottom caps) in reverse.

These have been done together, but that's confusing; it's a
single setting with several obscure effects.  Split it up.