Fix Roulette and Random. Both of these were picking a song immediately but making the wheel spinning noises.
Random looks like it was written with two behaviors in mind. The first is that it acts like Roulette except that where you are in the wheel has no correlation with the song picked and it stops spinning immediately. The second (which seems to have been added later), is to pick a song directly, a la Portal but without giving away any information about it. I went with the former since it was easier (less to clean up).

Roulette feels horribly wrong to me but at least the basic mechanics are there again. I think the Select() logic is wonky.