Previously, BeginStage was called in SGameplay and FinishStage was called in each place gameplay exits.  This incremented the stage number in SSM, and was brittle: FinishStage was sprinkled in many odd places.  Do both in SGameplay.
This increments the stage number as soon as gameplay ends.  This only
affects evaluation.  Use the stage number from StageStats there, instead
of the current stage, and adjust branches.  During evaluation, the
game is on the upcoming stage and displaying old data for the last
stage, instead of the stage continuing through evaluation.

This also moves a bunch of scoring-related stuff out of SEvaluation,
to the end of gameplay, where it belongs.  Eval is almost const, and things
won't break like they would before if Eval is not used.