Okay so Glenn decided to revert some Frieza code. I'm not going to remove everything Glenn did, as there was some useful code.
However, the ones I agreed with were:

75d5b63 "if a Noteskin fails to be loaded by the time the notes need to be displayed, attempt to force the default theme on and throw errors in to the log rather than crashing SM."
Glenn: There was no indication in the commit of when this was actually happening; this is unnecessary logic that will do nothing but hide bugs.
AJ: Based on some of the other code I've seen in the StepMania codebase, I'd say having this code in GetMetric isn't a good idea. There's an assert for checking if the noteskin isn't empty. That should suffice.

e27e971 "Adds metric options to flash the lifebar when a note is hit"
AJ: There are many possible ways of doing this without having to add a new metric.
The most extreme example is to code your own life bar and mess with JudgmentMessageCommand.
I'm not sure if StreamDisplay would respond to the message or not, but there are multiple ways of getting at the Stream. (ultralight, to be released, contains an example)