sm5svn 0dc297b [Glenn Maynard]
"Previously, we didn't create MusicWheel sorts when the screen loads; it was delayed until the
screen begins.  This allowed preloading before the style is set.  However, it caused a
major skip when the screen starts, because creating the sorts is fairly expensive.

Refactor this.  Load the sorts when the screen is loaded, and then go through them
and quickly filter out unplayable data and handle other initialization on begin.
This is much faster, so we can more easily transition seamlessly from the previous
menu.

One known limitation: the sorts are created before we know the style, so we have to
guess.  This means if you're in a style like dance-solo or pump-halfdouble, sorts will
still be based on regular mode.  This affects the difficulty, top grade and roulette sorts."