Wrap the autokeysound chain with RageSoundReader_Extend, quite a simple fix, but the song now no longer stops scrolling when there are no more auto keysounds to play, which is what I've been trying to fix for a long time.
It's now possible to play BMS which the BGM track is a bit too short
that in previous versions, the notes stop scrolling because there's
nothing more to play and it's impossible to finish the level.

It's also now possible to play BMS file which does not have any
autokeysounds at all, and also it's now possible to edit BMS files
(but the editor is very un-intuitive, but at least you can now add
timing changes to songs loaded from BMS files, such as speed segments
and scroll segments [they will be saved as .ssc file by the way]).