Stop some beat 0 crash bugs in the editor.
I suspect that the removal of shifting timing changes at beat 0
will stay in regardless. It is the removal of converting measures
to stops/delays that may be fought against. That is fine.

The bugs attached to these issues, 415 and 525, will only be
resolved when enough devs agree this is a fair fix.