Two easy to read complete opensource Noteskins called exact and easy. Play around with them and ajust anything you want There are made to test and increase the easy ness of reading simfiles with certain notesins, In this case one for accuracy which is "exact" and one for cluster simfiles which is "easy"
Easy is made to be easy to see with simfiles that are clustered
for example an simfile with an high bpm and a load of notes coming at the same
time, with an load of noteskins which have nice effects and its hard to read
these paterns.
so basicly you can read the file with clustered patterns which some people
including me have trouble with.

Exact is to test the accuracy, increasing your scores with Flawless, so people
get higher scores on easier files.