Revert "No need for coin data."
This reverts commit 3739c36956676e1b9a4b73afca998a7df2199e52.

That should be the last of the coin mode things. I will not let gimping the
game like this slip by. If you have a problem, I suggest forking.

AM's broken business model has nothing to do with us. Gimping coin modes won't
magically stop piracy, it will lead to more fragmentation in the community
than there already is (and there's a lot), and hurt adoption.

StepMania itself is a free game, but it's entirely within users rights granted
by the license to sell it. This is what you agree to by committing to it. It's
not our business if some choose to use their powers for less-than-good. If you
want to stop your content from being used with pay mode, we should think up
another way than ripping the whole damn thing out.