Allow fstab to use persistent device names for cards
Previously you had to use /dev/sdx or /dev/sdx1 for the device in /etc/fstab,
but the order in which these are assigned is non-deterministic on modern
kernels.  Now, if you want a specific card, you can use /dev/disk/by-uuid,
or if you want a specific port, you can use /dev/disk/by-path.