fix visual glitch on autoplay rolls
When there were tap notes on the same row as the start of a roll, autoplaying
the roll would result in the receptors on other tracks to glow.  This was
because the Step() call would use the roll's head row for a parameter, which
included the other notes.  This changes that parameter to -1 to automatically
use the current row in the song, which makes more sense anyway.

Fixes bug 241.