fix duplicate removal in editor help
The sort was in place for this purpose.  Let's use it on the modifier keys, but
not sort the regular keys, since the order matters for those.

This isn't really the right way to do this from the start, but that'll be
something that could be rewritten at another time given that our mappings are
still pretty much hardcoded.