fix ActorFrame diffusing and glowing children
When an ActorFrame's parent has a diffuse or glow, the ActorFrame uses
the static tempState in Actor::PreDraw(); When the ActorFrame draw's
it's children, those children also use tempState because their parent
has a diffuse( making their internalDiffuse not 1,1,1,1 ). If any child
is diffused it'll will change tempState, which will affect the
internalDiffuse given to the next child, and so on.

By determining the diffuse and glow to be passed to the children before
drawing any of them, any changes to tempState will not affect the
diffuse of subsequent children.