make ThemePrefsRows system actually save Prefs
This adds one function call to _fallback/Scripts/02 ThemePrefsRows.lua:  the call that saves the preferences.  The only reason ThemePrefs had been working in default theme (as far as I can tell) is because the same function was called from ScreenTitleMenu decorations.

That particular instance can be removed or left alone; this change does not interfere with it.  This change does allow other themers to use ThemePrefs and ThemePrefRows without figuring that out.