Allow OnlyShowActiveSection wheel to include current section title
IIDX hides inactive sections, but unlike the current behavior of our
OnlyShowActiveSection metric (based on certain Pumps according to
DaisukeMaster's comment), it retains the section title so that the
section can be closed normally.

New metric: MusicWheel.HideActiveSectionTitle, defaulting to true in
_fallback.  A bit unfortunate, but keeps backward compatibility with the
existing behavior of OnlyShowActiveSection.

If neither metric is true, wheel displays all titles plus the contents
of the active section (if any).

If OSAS is true but HAST is false, wheel displays the active section
title plus its contents, or all titles if no section is active.

If both are true, wheel displays only the active section's contents, or
all titles if no section is active.