log OS X version properly
There was code in place to log the OS X kernel version (14.1.0, for example) to log, but I can't recall if it has ever worked.  "Mac OS X unknown." has been logged for the last several releases.

This commit swaps out that C++ code for a little Obj-C which logs the OS X release version (10.10.2, for example) correctly.  In order to mix Obj-C with an otherwise C++ file, I changed the file extension from .cpp to .mm to make it Objective-C++.

I removed a now-unnecessary header (CoreServices.h) and added an import for Foundation.h