Introduce a fake judgment attribute.
Previously, only tap notes could be considered fake without the use of other timing segments. That has been changed so that any type of note can be fake.

This attribute is set up using =KV= format, where K is a key and V is a value. This approach is designed to allow for future simple attributes to be introduced in the future.

The editor can be used to introduce the judgments. Use `Shift + N` or `Shift + M` to cycle through the judgment types when placing notes, or use the Enter menu to modify judgments on that row.

A future pull request may attempt for Bonus (cannot miss) and Evil (cannot hit) attributes. At this time, stick to what works.

This PR may be cause for causing a minor version bump. This can break any noteskin that is dependent on displaying tap fakes distinctly from tap notes. Furthermore, a clean way of saving to the old .sm format with tap fakes has not been addressed yet.