Introduce GoogleTest and start unit testing.
The RageVectors were moved to a new project for easier testing.

Accessing/writing to the fields is now consistently done in one way.

Operating overloading is in proper effect. RageVectors can now be compared against one another.

The friend functions were removed: they are unneeded. Chain the operators the correct way.

More work will be done on this pending travis's successful build. There should be a way to make travis run the unit tests now that I think about it.