fix logic for attacks in gameplay disqualifying
When disqualification was turned on via Preferences, the old logic exhibited two problems:

1. charts with attacks were always disqualified from ranking
2. charts without attacks could be disqualified from ranking if the attack modifier was set to "Off"

So, this commit fixes those problems.  If a chart has attacks, and the attack modifier is used to turn them "Off", then disqualification is enforced.

Additionally, disqualification is enforced if the attack modifier is set to "Random Attacks."  This part might not be necessary, but I acted on the assumption that random attacks could theoretically make a chart easier.