Log with std::string.
A new header only library was brought in for assistance. This library allows enum classes to be treated as bit flags. No submodules this time: just one file by [Yuri Yaryshev](http://programmers.stackexchange.com/a/205567)

Some of the loggings themselves had to change due to the standards followed by cppformat. In short, it follows the POSIX guidelines more strictly than others. This means:

* No using `%p` to get the hex address of an object.
* No using enumeration values without casting them to a standard type first.

`RageException` was given similar treatment since it was easy to knock out at the same time.

Note to future enum class adders: do not use the word `None` as a value. It will conflict with X11.